refactor(AVP): decompose ARGuideViewModel into child ViewModels
Extract AIPerformance, PracticeFlow, Recording, and VirtualPianoPlacement
logic into dedicated child ViewModels. Add missing ARKit import for
world anchor tracking access.
